llm_server/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
